separated feedback component into a separate file for better readability
