+++ src/components/DefogDynamicViz.jsx
@@ -1,13 +1,5 @@
 import React, { useState, useContext, Fragment } from "react";
-import {
-  Button,
-  message,
-  Modal,
-  Input,
-  ConfigProvider,
-  Collapse,
-  Alert,
-} from "antd";
+import { Modal, ConfigProvider, Collapse, Alert, message } from "antd";
 import { CloseOutlined } from "@ant-design/icons";
 import { isEmpty, sentenceCase } from "./common/utils";
 import { styled } from "styled-components";
@@ -15,6 +7,7 @@ import ThumbsUp from "./svg/ThumbsUp";
 import ThumbsDown from "./svg/ThumbsDown";
 import { ThemeContext } from "../context/ThemeContext";
 import { TableChart } from "./TableChart";
+import Feedback from "./Feedback";
 // import Search from "antd/lib/input/Search";
 // import { BsPlusCircle } from "react-icons/bs";
 
@@ -37,17 +30,7 @@ const DefogDynamicViz = ({
 }) => {
   const { theme } = useContext(ThemeContext);
   const [modalVisible, setModalVisible] = useState(false);
-  const [hasReflected, setHasReflected] = useState(false);
-  const [reflectionFeedback, setReflectionFeedback] = useState("");
-  const [reflectionColDescriptions, setReflectionColDescriptions] = useState(
-    [],
-  );
   const [feedbackType, setFeedbackType] = useState(null);
-  const [reflectionRefQueries, setReflectionRefQueries] = useState([]);
-  const [reflectionLoading, setReflectionLoading] = useState(false);
-  const [glossary, setGlossary] = useState("");
-  const [postReflectionLoading, setPostReflectionLoading] = useState(false);
-  const { TextArea } = Input;
 
   const warningMessage = (
     <Alert
@@ -57,6 +40,23 @@ const DefogDynamicViz = ({
       showIcon
     />
   );
+
+  const uploadPositiveFeedback = async () => {
+    await fetch(`https://api.defog.ai/feedback`, {
+      method: "POST",
+      headers: {
+        "Content-Type": "application/json",
+      },
+      body: JSON.stringify({
+        apiKey: apiKey,
+        response: response,
+        feedback: feedbackType,
+        dev: dev,
+      }),
+    });
+
+    message.info("Thank you for your feedback!");
+  };
 
   // if no response, return error
   if (!response || isEmpty(response)) {
@@ -68,182 +68,6 @@ const DefogDynamicViz = ({
       </ErrorMessageWrap>
     );
   }
-
-  const uploadFeedback = async (
-    feedback,
-    feedbackText = "",
-    giveSuggestions = guidedTeaching,
-  ) => {
-    if (feedback === "Good") {
-      setFeedbackType("Good");
-      await fetch(`https://api.defog.ai/feedback`, {
-        method: "POST",
-        headers: {
-          "Content-Type": "application/json",
-        },
-        body: JSON.stringify({
-          apiKey: apiKey,
-          response: response,
-          feedback: feedback,
-          dev: dev,
-        }),
-      });
-
-      message.info("Thank you for your feedback!");
-    } else {
-      setFeedbackType("Bad");
-      // send feedback over to the server
-      await fetch(`https://api.defog.ai/feedback`, {
-        method: "POST",
-        headers: {
-          "Content-Type": "application/json",
-        },
-        body: JSON.stringify({
-          apiKey: apiKey,
-          response: response,
-          feedback: feedback,
-          text: feedbackText,
-          dev: dev,
-        }),
-      });
-
-      if (giveSuggestions) {
-        // send the error to the reflect endpoint
-        setReflectionLoading(true);
-        message.info(
-          "Preparing improved instruction sets for the model. This can take up to 30 seconds. Thank you for your patience.",
-        );
-
-        // first, get the metadata so that we can easily compare it against the reflection
-        const metadataResp = await fetch(`https://api.defog.ai/get_metadata`, {
-          method: "POST",
-          headers: {
-            "Content-Type": "application/json",
-          },
-          body: JSON.stringify({
-            api_key: apiKey,
-            dev: dev,
-          }),
-        });
-        const { table_metadata, glossary } = await metadataResp.json();
-        setGlossary(glossary);
-
-        const reflectResp = await fetch(
-          `https://api.defog.ai/reflect_on_error`,
-          {
-            method: "POST",
-            headers: {
-              "Content-Type": "application/json",
-            },
-            body: JSON.stringify({
-              api_key: apiKey,
-              question: response.question,
-              sql_generated: response.generatedSql,
-              error: feedbackText,
-              dev: dev,
-            }),
-          },
-        );
-
-        const {
-          feedback,
-          instruction_set,
-          column_descriptions,
-          reference_queries,
-        } = await reflectResp.json();
-
-        // table_metadata is currently an object in the form of {table_name: [{column_name: ..., description: ...}]}
-        // we need to convert it to an array of objects in the form of {table_name: ..., column_name: ..., description: ...}
-        const column_descriptions_array = [];
-        for (const table_name in table_metadata) {
-          const columns = table_metadata[table_name];
-          columns.forEach((column) => {
-            column_descriptions_array.push({
-              table_name: table_name,
-              column_name: column.column_name,
-              data_type: column.data_type,
-              original_description: column.column_description,
-            });
-          });
-        }
-
-        // add a new key to each item in column_descriptions called "original description". this is the column description from the metadata
-        const updatedDescriptions = [];
-
-        column_descriptions_array.forEach((item) => {
-          const found = column_descriptions.find(
-            (meta) =>
-              meta.table_name === item.table_name &&
-              meta.column_name === item.column_name,
-          );
-          if (found) {
-            updatedDescriptions.push({
-              ...item,
-              updated_description: found.description,
-            });
-          }
-        });
-
-        setHasReflected(true);
-        setReflectionFeedback(feedback);
-        setGlossary(glossary + "\n\n(new instructions)\n\n" + instruction_set);
-        setReflectionColDescriptions(updatedDescriptions);
-        setReflectionRefQueries(reference_queries);
-        setReflectionLoading(false);
-      } else {
-        setModalVisible(false);
-      }
-    }
-  };
-
-  const updateNewInstructions = async () => {
-    setPostReflectionLoading(true);
-    // update glossary
-    await fetch(`https://api.defog.ai/update_glossary`, {
-      method: "POST",
-      headers: {
-        "Content-Type": "application/json",
-      },
-      body: JSON.stringify({
-        api_key: apiKey,
-        glossary: glossary,
-        dev: dev,
-      }),
-    });
-
-    // update golden queries
-    await fetch(`https://api.defog.ai/update_golden_queries`, {
-      method: "POST",
-      headers: {
-        "Content-Type": "application/json",
-      },
-      body: JSON.stringify({
-        api_key: apiKey,
-        golden_queries: reflectionRefQueries,
-        scrub: false,
-        dev: dev,
-      }),
-    });
-
-    // update column descriptions
-    await fetch(`https://api.defog.ai/update_column_descriptions`, {
-      method: "POST",
-      headers: {
-        "Content-Type": "application/json",
-      },
-      body: JSON.stringify({
-        api_key: apiKey,
-        column_descriptions: reflectionColDescriptions,
-        dev: dev,
-      }),
-    });
-
-    setPostReflectionLoading(false);
-    setModalVisible(false);
-    message.info(
-      "The model's instruction set has now been updated. Thank you for the feedback!",
-    );
-  };
 
   let results;
 
@@ -263,10 +87,20 @@ const DefogDynamicViz = ({
 
         <FeedbackWrap theme={theme.config}>
           <p>How did we do with is this query?</p>
-          <button onClick={() => uploadFeedback("Good")}>
+          <button
+            onClick={() => {
+              setFeedbackType("Good");
+              uploadPositiveFeedback();
+            }}
+          >
             <ThumbsUp />
           </button>
-          <button onClick={() => setModalVisible(true)}>
+          <button
+            onClick={() => {
+              setFeedbackType("Bad");
+              setModalVisible(true);
+            }}
+          >
             <ThumbsDown />
           </button>
         </FeedbackWrap>
@@ -375,10 +209,20 @@ const DefogDynamicViz = ({
 
             <FeedbackWrap theme={theme.config}>
               <p>How did we do with is this query?</p>
-              <button onClick={() => uploadFeedback("Good")}>
+              <button
+                onClick={() => {
+                  setFeedbackType("Good");
+                  uploadPositiveFeedback();
+                }}
+              >
                 <ThumbsUp />
               </button>
-              <button onClick={() => setModalVisible(true)}>
+              <button
+                onClick={() => {
+                  setFeedbackType("Bad");
+                  setModalVisible(true);
+                }}
+              >
                 <ThumbsDown />
               </button>
             </FeedbackWrap>
@@ -411,126 +255,14 @@ const DefogDynamicViz = ({
           width={800}
         >
           <FeedbackModalWrap theme={theme.config}>
-            <>
-              <TextArea
-                rows={4}
-                className="feedback-text"
-                placeholder="Optional"
-                id={"feedback-text-" + questionId}
-              />
-              {!hasReflected ? (
-                <>
-                  <Button
-                    loading={reflectionLoading}
-                    disabled={reflectionLoading}
-                    onClick={() => {
-                      // upload feedback, based on the value of the text area
-                      const feedbackText = document.getElementById(
-                        "feedback-text-" + questionId,
-                      ).value;
-                      uploadFeedback("Bad", feedbackText, false);
-                    }}
-                  >
-                    Submit
-                  </Button>
-                  {guidedTeaching ? (
-                    <Button
-                      loading={reflectionLoading}
-                      disabled={reflectionLoading}
-                      onClick={() => {
-                        // upload feedback, based on the value of the text area
-                        const feedbackText = document.getElementById(
-                          "feedback-text-" + questionId,
-                        ).value;
-                        uploadFeedback("Bad", feedbackText, guidedTeaching);
-                      }}
-                    >
-                      Submit and get suggestions for improvement
-                    </Button>
-                  ) : null}
-                </>
-              ) : (
-                <>
-                  <p>{reflectionFeedback}</p>
-
-                  <p>Instruction Set:</p>
-                  <TextArea
-                    rows={8}
-                    value={glossary}
-                    onChange={(e) => setGlossary(e.target.value)}
-                    style={{
-                      marginTop: "1em",
-                      marginBottom: "1em",
-                    }}
-                  />
-                  <p>Column Descriptions:</p>
-                  <ul>
-                    {reflectionColDescriptions.map((item, idx) => {
-                      return (
-                        <li key={idx}>
-                          Table Name: {item.table_name}
-                          <br />
-                          Column Name: {item.column_name}
-                          <br />
-                          Original Description: {item.original_description}
-                          <br />
-                          Suggested Description:{" "}
-                          <TextArea
-                            rows={2}
-                            value={item.updated_description}
-                            onChange={(e) => {
-                              const updatedDescriptions = [
-                                ...reflectionColDescriptions,
-                              ];
-                              updatedDescriptions[idx].updated_description =
-                                e.target.value;
-                              setReflectionColDescriptions(updatedDescriptions);
-                            }}
-                            style={{
-                              marginBottom: "1em",
-                            }}
-                          />
-                        </li>
-                      );
-                    })}
-                  </ul>
-                  <p>Reference Queries:</p>
-                  <ul>
-                    {reflectionRefQueries.map((item, idx) => {
-                      return (
-                        <li key={idx}>
-                          Question: {item.question}
-                          <br />
-                          SQL:{" "}
-                          <TextArea
-                            rows={8}
-                            value={item.sql}
-                            onChange={(e) => {
-                              const updatedQueries = [...reflectionRefQueries];
-                              updatedQueries[idx].sql = e.target.value;
-                              setReflectionRefQueries(updatedQueries);
-                            }}
-                            style={{
-                              marginBottom: "1em",
-                            }}
-                          />
-                        </li>
-                      );
-                    })}
-                  </ul>
-
-                  <Button
-                    onClick={() => {
-                      updateNewInstructions();
-                    }}
-                    loading={postReflectionLoading}
-                    disabled={postReflectionLoading}
-                  >
-                    Update Instructions
-                  </Button>
-                </>
-              )}
-            </>
+            <Feedback
+              apiKey={apiKey}
+              dev={dev}
+              guidedTeaching={guidedTeaching}
+              questionId={questionId}
+              response={response}
+              setModalVisible={setModalVisible}
+            />
           </FeedbackModalWrap>
         </Modal>
       </ConfigProvider>
